Remade remake file
The remake now handles installation a lot better.

Full commit message list:
* initial start of remaking makefile - still need to handle testing and linting

* small fix for handling config files

* added in linting and initial testing to makefile

* added ext build and lodrc setup

* some PR feedback, still need to work on if venv should be included or not

* removed management of venv, now leaves it up to the user
