+++ Makefile
@@ -1,67 +1,60 @@
-PYTHONFILES = $(wildcard \
-	learning_observer/learning_observer/*py \
-	learning_observer/util/*py \
-	gitserve/gitserve/*py \
-	gitserve/*py \
-	learning_observer/learning_observer/pubsub/*py \
-	learning_observer/learning_observer/stream_analytics/*py \
-)
+PACKAGES ?= wo,awe
 
 run:
 	# If you haven't done so yet, run: make install
-	# Also, run: workon learning_observer
-	# The export PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION makes this slower, but avoids
-	# annoying version issues in dev. This should be resolved at some point.
-	export PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python; cd learning_observer && python learning_observer
+	# we need to make sure we are on the virtual env when we do this
+	cd learning_observer && python learning_observer
 
-# Build browser extension
-extension-package:
-	# I won't build this for you, but I will give you instructions, since
-	# the set of options will vary.
-	#
-	# On my last system, I needed:
-	#	google-chrome --pack-extension=extension --pack-extension-key=extension.pem --disable-setuid-sandbox --no-gpu --no-sandbox --headless
-	# On my current system
-	#	google-chrome --pack-extension=extension --pack-extension-key=extension.pem
-	# The pem file is your private key. You'll need to make one.
-	# The output will be called extension.crx
+venv:
+	pip install --no-cache-dir -r requirements.txt
 
-codestyle:
-	# Check code style quality
-	#
-	# In pycodestyle, we ignore E501 (line too long) and W503 (which
-	# requires a choice.
-        # https://stackoverflow.com/questions/57074300/what-is-the-recommended-way-to-break-long-if-statement-w504-line-break-after-b
-	#
-	# In pylint, we ignore:
-	# 1. W0613: unused arguments (common for e.g. `request` parameter)
-	# 2. E501/C0301: line too long (obsolete with modern computers
-	#    ref: https://lkml.org/lkml/2020/5/29/1038)
-	#    We still aim for shorter lines, but it's not a showstopper if
-	#    we break the limit occasionally. We may re-examine this decision
-	#    if we have sufficient developers working on VT100 terminals.
-	# 3. W0511: TODO: We're gonna have a lot of these, and we want to
-	#    encourage leaving these around. As a coding style, we want to
-	#    get interface and overall structures right, and then clean
-	#    up e.g. performance/scaling, exception handling, test coverage,
-	#    etc. once we know what we're doing.
-	# 4. R0913: Too many arguments. That's a relic of web apps. Sorry.
-	#
-	# Generally, pycodestyle issues are showstopppers for pushing to
-	# upstream. Pylint issues are worth an occasional cleanup pass, but we
-	# can tolerate.
+# install commands
+install: venv
+	# The following only works with specified packages
+	# we need to install learning_observer in dev mode to
+	# more easily pass in specific files we need, such as creds
+	pip install --no-cache-dir -e learning_observer/[${PACKAGES}]
+	# TODO resolve the lodrc-current symlink and fetch that url instead
+	@LODRC_CURRENT=$$(curl -s https://raw.githubusercontent.com/ETS-Next-Gen/lo_assets/main/lo_dash_react_components/lo_dash_react_components-current.tar.gz); \
+	pip install https://raw.githubusercontent.com/ETS-Next-Gen/lo_assets/main/lo_dash_react_components/$${LODRC_CURRENT}
 
-	pycodestyle --ignore=E501,W503 $(PYTHONFILES)
-	pylint -d W0613,W0511,C0301,R0913,too-few-public-methods $(PYTHONFILES)
+install-dev: venv
+	# TODO create a dev requirements file
+	pip install --no-cache-dir -e learning_observer/[${PACKAGES}]
+	. ${HOME}/.nvm/nvm.sh && nvm use && pip install -v -e modules/lo_dash_react_components/
 
-install:
-	# Run:
-	#    mkvirtualenv learning_observer
-	#    pip install -r requirements.txt
-	#    cd learning_observer
-	#    python setup.py develop
-	#    python learning_observer
-	# And then clean up this Makefile to do it for you automatically :)
+# testing commands
+test:
+	# this is where we run doctests
+	pytest modules/wo_highlight_dashboard
 
-sphinx:
-	cd learning_observer/docs; make html
+# Linting commands
+linting-setup:
+	# Setting up linting related packages
+	pip install pycodestyle pylint
+	npm install
+
+linting-python:
+	# Linting Python modules
+	pycodestyle --ignore=E501,W503 $$(git ls-files 'learning_observer/*.py' 'modules/*.py')
+	pylint -d W0613,W0511,C0301,R0913,too-few-public-methods $$(git ls-files 'learning_observer/*.py' 'modules/*.py')
+
+linting-node:
+	# TODO each of these have lots of errors and block
+	# the next item from running
+	# Starting to lint Node modules
+	# Linting Javascript
+	npm run lint:js
+	# Linting CSS
+	npm run lint:css
+	# Finding any unused CSS files
+	npm run find-unused-css
+
+linting: linting-setup linting-python linting-node
+	# Finished linting
+
+build-writing-ext:
+	# Installing LO Event
+	cd modules/lo_event && npm install & npm link lo_event
+	# Building extension
+	cd extension/writing-process && npm install && npm run build
