+++ learning_observer/learning_observer/creds.yaml.example
@@ -52,8 +52,11 @@ kvs:
   default:
     type: stub
     expiry: 6000
+  memoization:
+    type: redis_ephemeral
+    expiry: 60
 roster_data:
-    source: filesystem  # Can be set to google_api, all, test, or filesystem
+    source: all  # Can be set to google_api, all, test, or filesystem
 aio:  # User session; used for log-ins.
     session_secret: {unique-aio-session-key}  # This should be a unique secret key for YOUR deployment
     session_max_age: 3600  # In seconds. This may be short for auth dev (e.g. <5 mins), intermediate for deploy (a few hours?), and long for e.g. testing other parts of the system (or set to null, for lifetime of the browser)
