+++ learning_observer/setup.py
@@ -5,13 +5,8 @@ To set up locally for development, run `python setup.py develop`, in a
 virtualenv, preferably.
 '''
 
-from setuptools import setup
-from setuptools.command.develop import develop as _develop
-from setuptools.command.install import install as _install
-
+from setuptools import setup, find_packages
 import os
-import subprocess
-import sys
 
 my_path = os.path.dirname(os.path.realpath(__file__))
 parent_path = os.path.abspath(os.path.join(my_path, os.pardir))
@@ -25,48 +20,12 @@ def clean_requirements(filename):
     requirements = [s.strip() for s in open(file_path).readlines() if len(s) > 1]
     return requirements
 
-
-new_path = '/'.join(sys.executable.split('/')[:-1])
-current_path = os.environ['PATH']
-modified_env = {'PATH': f'{new_path}{os.pathsep}{current_path}'}
-
-
-class LOInstall(_develop):
-    '''
-    LO_dash_react_components relies on Dash to be installed.
-    To ensure Dash (and any other downstream dependencies) are resolved first,
-    we install LODRC after the install_requires
-    '''
-    def run(self):
-        _develop.run(self)
-        result = subprocess.run(
-            ['pip', 'install', '-v', '-e', 'git+https://github.com/ETS-Next-Gen/writing_observer.git#egg=lo_dash_react_components&subdirectory=modules/lo_dash_react_components'],
-            env=modified_env, stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-
-
-class LOInstallTEST(_install):
-    '''
-    LO_dash_react_components relies on Dash to be installed.
-    To ensure Dash (and any other downstream dependencies) are resolved first,
-    we install LODRC after the install_requires
-    '''
-    def run(self):
-        _install.run(self)
-        result = subprocess.run(
-            ['pip', 'install', '-v', '-e', 'git+https://github.com/ETS-Next-Gen/writing_observer.git#egg=lo_dash_react_components&subdirectory=modules/lo_dash_react_components'],
-            env=modified_env, stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-
-
 setup(
     install_requires=clean_requirements(req_path),
     extras_require={
         "wo": clean_requirements(wo_path),
         "awe": clean_requirements(awe_path)
     },
-    cmdclass={
-        'develop': LOInstall,
-        'install': LOInstallTEST
-    }
+    packages=find_packages(),
+    package_data={'': ['util/*', 'static/**/*', 'static_data/*.template', 'creds.yaml.example', 'communication_protocol/schema.json']}
 )
